+++ src/main/java/blackjack/domain/card/Deck.java
@@ -4,7 +4,11 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+import static blackjack.domain.Game.WINNING_NUMBER;
+
 public class Deck {
+
+    private static final int ACE_EXCEPTION_NUMBER = 10;
 
     private final List<Card> deck;
 
@@ -22,5 +26,25 @@ public class Deck {
 
     public List<Card> getCards() {
         return Collections.unmodifiableList(deck);
+    }
+
+    public int getScore() {
+        int scoreExceptAce = deck.stream()
+                .filter(card -> !card.isAce())
+                .mapToInt(Card::getScore)
+                .sum();
+
+        return deck.stream()
+                .filter(Card::isAce)
+                .mapToInt(card -> scoreCalculate(scoreExceptAce, card))
+                .sum() + scoreExceptAce;
+    }
+
+    private int scoreCalculate(int scoreExceptAce, Card card) {
+        if (scoreExceptAce + card.getScore() + ACE_EXCEPTION_NUMBER > WINNING_NUMBER) {
+            return card.getScore();
+        }
+
+        return card.getScore() + ACE_EXCEPTION_NUMBER;
     }
 }
