+++ src/main/java/blackjack/domain/player/Gamer.java
@@ -1,17 +1,11 @@
 package blackjack.domain.player;
 
 import blackjack.domain.Status;
-import blackjack.util.BlackjackScoreCalculator;
-import blackjack.util.ScoreCalculator;
 
 public class Gamer extends Player {
 
-    public Gamer(String name, ScoreCalculator scoreCalculator) {
-        super(name, scoreCalculator);
-    }
-
     public Gamer(String name) {
-        this(name, new BlackjackScoreCalculator());
+        super(name);
     }
 
     @Override
