+++ src/main/java/blackjack/domain/player/Player.java
@@ -4,7 +4,6 @@ import blackjack.domain.Status;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Deck;
 import blackjack.exception.CardDuplicateException;
-import blackjack.util.ScoreCalculator;
 
 import java.util.List;
 
@@ -13,12 +12,9 @@ public abstract class Player {
     private final Deck deck;
     private final String name;
 
-    private final ScoreCalculator scoreCalculator;
-
-    public Player(String name, ScoreCalculator scoreCalculator) {
+    public Player(String name) {
         this.deck = new Deck();
         this.name = name;
-        this.scoreCalculator = scoreCalculator;
     }
 
     public void addCardToDeck(Card card) {
@@ -34,7 +30,7 @@ public abstract class Player {
     }
 
     public int getScore() {
-        return scoreCalculator.apply(deck);
+        return deck.getScore();
     }
 
     public Status getStatus() {
