--- src/main/java/blackjack/util/ScoreCalculator.java
@@ -1,8 +1,0 @@
-package blackjack.util;
-
-import blackjack.domain.card.Deck;
-
-public interface ScoreCalculator {
-
-    int apply(Deck deck);
-}
